move helper functions to pace.util.testing
