+++ pace-util/pace/util/testing/__init__.py
@@ -1,2 +1,3 @@
 from . import gt4py_cupy, gt4py_numpy
+from .comparison import compare_arr, compare_scalar, success, success_array
 from .dummy_comm import ConcurrencyError, DummyComm
